Feature/backend api

+++ src/common/image/unlinkImage.js
@@ -1,20 +1,26 @@
 const fs = require("fs");
+const path = require("path");
 
-function unlinkImage(imagePaths) {
-  if (typeof imagePaths === "string") {
-    // If it's a single image path, convert it to an array
-    imagePaths = [imagePaths];
+function unlinkImage(imagePath) {
+  if (!/^\/?public\//.test(imagePath)) {
+    // If not, prepend "public/" to the path
+    imagePath = path.join("public/", imagePath);
+}
+
+  const fileExists = fs.existsSync(imagePath);
+  if (!fileExists) {
+    console.log(`File ${imagePath} does not exist`);
+    return;
   }
 
-  imagePaths.forEach((imagePath) => {
-    fs.unlink(imagePath, (err) => {
-      if (err) {
-        console.error(`Error deleting the file ${imagePath}:`, err);
-      } else {
-        console.log(`File ${imagePath} deleted successfully`);
-      }
-    });
+  fs.unlink(imagePath, (err) => {
+    if (err) {
+      console.error(`Error deleting the file ${imagePath}:`, err);
+    } else {
+      console.log(`File ${imagePath} deleted successfully`);
+    }
   });
 }
 
-module.exports = unlinkImage
+module.exports = unlinkImage;
+
